make it better

+++ authApp/views/userDetailView.py
@@ -1,6 +1,6 @@
 from authProject import settings
 
-from rest_framework import generics, status
+from rest_framework import generics, status, views
 from rest_framework.response import Response
 from rest_framework_simplejwt.backends import TokenBackend
 
@@ -8,10 +8,8 @@ from authApp.models.user import User
 from authApp.serializers.userSerializer import UserSerializer
 from authApp.serializers.userUpdateSerializer import UserUpdateSerializer
 
-class UserDetailView(generics.RetrieveAPIView):
-  # users = User.objects.all()
-  queryset = User.objects.all()
-  serializer_class = UserSerializer
+class UserDetailView(views.APIView):
+  users = User.objects.all()
   # serializer = UserSerializer(users, many=True)
  
   # def get(self, request):
@@ -26,27 +24,35 @@ class UserDetailView(generics.RetrieveAPIView):
       error = { 'message': 'Unathorized user!' }
       return Response(error, status=status.HTTP_401_UNAUTHORIZED)
     
-    # user = self.users.get(id=kwargs['pk'])
-    # user_serializer = UserSerializer(user)
+    user = self.users.get(id=kwargs['pk'])
+    user_serializer = UserSerializer(user)
 
-    # return Response(user_serializer.data, status=status.HTTP_200_OK)
+    return Response(user_serializer.data, status=status.HTTP_200_OK)
 
-    return super().get(request, *args, **kwargs)
 
   def put(self, request, *args, **kwargs):
-    user = User.objects.get(id=kwargs['pk'])
-    user_serializer = UserUpdateSerializer(user, data=request.data)
+    user = self.users.filter(id=kwargs['pk']).first() # Similar a self.users.get(id=kwargs['pk']) pero evita errores en caso de que el usuario no ea encontrado
 
-    if (user_serializer.is_valid()): 
-      user_serializer.save()
-      return Response({ 'message': 'Usuario actualizado' }, status=status.HTTP_200_OK)
+    if (user):
+      # Si le pasan como primer parámetro la instancia al serializador, ya pueden llamar el método update, si no, les pedirá que creen un método 
+      user_serializer = UserUpdateSerializer(user, data=request.data)
 
-    else:
+      if (user_serializer.is_valid(raise_exception=True)):
+        user_serializer.save()
+       
+        return Response({ 'message': 'Usuario actualizado' }, status=status.HTTP_200_OK)
+
       return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    return Response({ 'message': 'Usuario no encontrado' }, status=status.HTTP_404_NOT_FOUND) 
 
 
   def delete(self, request, *args, **kwargs):
-    user = User.objects.get(id=kwargs['pk'])
-    user.delete()
+    user = self.users.filter(id=kwargs['pk']).first()
+    
+    if (user):
+      user.delete()
 
-    return Response({ 'message': 'Usuario eliminado' }, status=status.HTTP_200_OK)
+      return Response({ 'message': 'Usuario eliminado' }, status=status.HTTP_200_OK)
+
+    return Response({ 'message': 'Usuario no encontrado' }, status=status.HTTP_404_NOT_FOUND)
